Draft serialization API (types, data, constants, IR);
