+++ sigma-ser/src/serializer.rs
@@ -8,6 +8,9 @@ pub enum SerializationError {
     /// Failed to parse op
     #[error("op parsing error")]
     InvalidOpCode,
+    /// Lacking support for the op
+    #[error("not implemented op error")]
+    NotImplementedOpCode(u8),
     /// Failed to parse type
     #[error("type parsing error")]
     InvalidTypePrefix,
