+++ sigma-tree/src/lib.rs
@@ -14,6 +14,7 @@ mod ast;
 mod data;
 mod ergo_tree;
 mod eval;
+mod serialization;
 mod types;
 
 pub use ergo_tree::*;
